+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 ### >>> CHANGELOG <<<
 
 ### Version 0.2.0:
+- Added TlgPolygon class and example
+- Added TlgVideo.LoadFromFile/LoadFromZipFile
+- Added TlgGame class, lgRunGame routine
+- Renamed TlgConsole to TlgTerminal
 - Added TlgTexture.Lock/Unlock/GetPixel/SetPixel()
 - Added TlgWindow.GetPixel/SetPixel()
-- Added pocketlang
+- Added pocketlang scripting
 - Added first draft docs (LGT.chm, LGT.pdf)
 - Add a Sleep(0) to the while loop in TlgTaskList.Start()
 - Update TlgTexture.DrawTiled(), replace  AViewportWidth, AViewportHeight with AWindow: TlgWindow and get the viewport from